ADD database caching

+++ app/src/main/java/com/lamantin/sildingpanelayoutdemo/models/api/Api.java
@@ -14,7 +14,7 @@ public interface Api {
     @GET("/users/{user}/albums")
     Observable<List<AlbumDTO>> getAlbumsByUser(@Path("user") int user);
     @GET("/albums/{album}/photos")
-    Observable<List<PhotoDTO>> getPhotosByAlbum(@Path("album") int album);
+    Observable<List<PhotoDTO>> getPhotosByAlbum(@Path("album") long album);
     @GET("/albums/11/photos")
     Observable<List<PhotoDTO>> getPhotoHistory();
 }
